sm_modified_cnn/requirements.txt
nltk==3.2.1
numpy==1.11.3
pytorch==0.1.12
